Crate bdk_testenv

Re-exports

pub use electrsd;

pub use electrsd::bitcoind;

pub use electrsd::bitcoind::anyhow;

pub use electrsd::bitcoind::bitcoincore_rpc;

pub use electrsd::electrum_client;

Modules

utils

Macros

block_id

chain_update

changeset

hash

local_chain

Structs

Config

Configuration parameters.

TestEnv

Struct for running a regtest environment with a single bitcoind node with an electrs instance connected to it.