Struct bdk_testenv::TestEnv

pub struct TestEnv {
    pub bitcoind: BitcoinD,
    pub electrsd: ElectrsD,
}

Struct for running a regtest environment with a single bitcoind node with an electrs instance connected to it.

Fields
bitcoind: BitcoinDelectrsd: ElectrsD
Implementations

impl TestEnv

pub fn new() -> Result<Self>

Construct a new TestEnv instance with the default configuration used by BDK.

pub fn new_with_config(config: Config<'_>) -> Result<Self>

Construct a new TestEnv instance with the provided Config.

pub fn electrum_client(&self) -> &impl ElectrumApi

Exposes the [ElectrumApi] calls from the Electrum client.

pub fn rpc_client(&self) -> &impl RpcApi

Exposes the [RpcApi] calls from [bitcoincore_rpc].

pub fn reset_electrsd(self) -> Result<Self>

pub fn mine_blocks( &self, count: usize, address: Option<Address> ) -> Result<Vec<BlockHash>>

Mine a number of blocks of a given size count, which may be specified to a given coinbase address.

pub fn mine_empty_block(&self) -> Result<(usize, BlockHash)>

Mine a block that is guaranteed to be empty even with transactions in the mempool.

pub fn wait_until_electrum_sees_block(&self, timeout: Duration) -> Result<()>

This method waits for the Electrum notification indicating that a new block has been mined. timeout is the maximum Duration we want to wait for a response from Electrsd.

pub fn wait_until_electrum_sees_txid( &self, txid: Txid, timeout: Duration ) -> Result<()>

This method waits for Electrsd to see a transaction with given txid. timeout is the maximum Duration we want to wait for a response from Electrsd.

pub fn invalidate_blocks(&self, count: usize) -> Result<()>

Invalidate a number of blocks of a given size count.

pub fn reorg(&self, count: usize) -> Result<Vec<BlockHash>>

Reorg a number of blocks of a given size count. Refer to TestEnv::mine_empty_block for more information.

pub fn reorg_empty_blocks( &self, count: usize ) -> Result<Vec<(usize, BlockHash)>>

Reorg with a number of empty blocks of a given size count.

pub fn send( &self, address: &Address<NetworkChecked>, amount: Amount ) -> Result<Txid>

Send a tx of a given amount to a given address.

pub fn make_checkpoint_tip(&self) -> CheckPoint

Create a checkpoint linked list of all the blocks in the chain.

pub fn genesis_hash(&self) -> Result<BlockHash>

Get the genesis hash of the blockchain.